Add users and tasks creation end points (Lecture 89 & 90)
