+++ src/db/mongoose.js
@@ -1,80 +1,6 @@
 const mongoose = require('mongoose');
-const validator = require('validator');
 
 mongoose.connect('mongodb://127.0.0.1:27017/task-manager-api', {
     useNewUrlParser: true,
     useCreateIndex: true
-});
-
-const User = mongoose.model('User', {
-    name: {
-        type: String,
-        required: true,
-        trim: true
-    },
-    email: {
-        type: String,
-        required: true,
-        trim: true,
-        lowercase: true,
-        validate(value) {
-            if (!validator.isEmail(value)) {
-                throw new Error('Email is invalid!');
-            }
-        }
-    },
-    password: {
-        type: String,
-        required: true,
-        minlength: 7,
-        trim: true,
-        validate(value) {
-            if (value.toLowerCase().includes('password')) {
-                throw new Error ('Password cannot contain "password"');
-            }
-        }
-    },   
-    age: {
-        type: Number,
-        default: 0,
-        validate(value) {
-            if (value < 0) {
-                throw new Error('Age must be a positive number');
-            }
-        }
-    }
-});
-
-/*const me = new User({
-    name: '    Andres ',
-    email: 'MYEMAIL@mead.io   ',
-    password: 'phone098!'
-});
-
-me.save().then(() => {
-    console.log(me);
-}).catch((error) => {
-    console.log('Error: ', error);
-});*/
-
-const Task = mongoose.model('Task', {
-    description: {
-        type: String,
-        required: true,
-        trim: true
-    },
-    completed: {
-        type: Boolean,
-        default: false
-    }
-});
-
-const task = new Task({
-    description: '  Eat lunch'
-});
-
-task.save().then(() => {
-    console.log(task);
-}).catch((error) => {
-    console.log(error);
 });
